+++ src/hb/aat_layout.rs
@@ -3,7 +3,7 @@
 use super::buffer::{hb_buffer_t, hb_glyph_info_t};
 use super::hb_font_t;
 use super::hb_tag_t;
-use super::shape_plan::hb_ot_shape_plan_t;
+use super::ot_shape_plan::hb_ot_shape_plan_t;
 use super::{aat_layout_kerx_table, aat_layout_morx_table, aat_layout_trak_table};
 
 pub type hb_aat_layout_feature_type_t = u8;
