+++ src/hb/mod.rs
@@ -43,8 +43,8 @@ mod ot_shape_complex_use_table;
 mod ot_shape_complex_vowel_constraints;
 mod ot_shape_fallback;
 mod ot_shape_normalize;
+pub mod ot_shape_plan;
 pub mod shape;
-pub mod shape_plan;
 mod tag;
 mod tag_table;
 mod text_parser;
