+++ src/hb/ot_layout_gpos_table.rs
@@ -7,7 +7,7 @@ use super::hb_font_t;
 use super::ot_layout::*;
 use super::ot_layout_common::{lookup_flags, PositioningLookup, PositioningTable};
 use super::ot_layout_gsubgpos::{skipping_iterator_t, Apply, OT::hb_ot_apply_context_t};
-use super::shape_plan::hb_ot_shape_plan_t;
+use super::ot_shape_plan::hb_ot_shape_plan_t;
 use crate::Direction;
 
 pub fn position(plan: &hb_ot_shape_plan_t, face: &hb_font_t, buffer: &mut hb_buffer_t) {
